+++ activerecord/lib/active_record/associations/belongs_to_association.rb
@@ -25,20 +25,21 @@ module ActiveRecord
         when :destroy
           raise ActiveRecord::Rollback unless target.destroy
         when :destroy_async
-          primary_key_column = reflection.active_record_primary_key
-          ids = foreign_key.map { |col| owner.public_send(col) }
-
           association_class = if reflection.polymorphic?
-            owner.public_send(foreign_type)
+            target.class
           else
             reflection.klass
           end
+
+          primary_key_column = reflection.query_primary_key(association_class)
+          query_foreign_key = ActiveRecord::Key.for(reflection.query_foreign_key)
+          ids = query_foreign_key.map { |column| owner.public_send(column) }
 
           enqueue_destroy_association(
             owner_model_name: owner.class.to_s,
             owner_id: owner.id,
             association_class: association_class.to_s,
-            association_ids: foreign_key.composite? ? [ids] : ids,
+            association_ids: query_foreign_key.composite? ? [ids] : ids,
             association_primary_key_column: primary_key_column,
             ensuring_owner_was_method: options.fetch(:ensuring_owner_was, nil)
           )
@@ -141,7 +142,12 @@ module ActiveRecord
         end
 
         def replace_keys(record, force: false)
-          target_key_values = record ? ActiveRecord::Key.for(primary_key(record.class)).map { |col| record.read_attribute(col) } : []
+          target_key = if record
+            reflection.query_key_mapping(record.class).foreign_key_associated_record_columns
+          end
+          target_key = ActiveRecord::Key.for(target_key)
+
+          target_key_values = target_key.map { |key| record.read_attribute(key) }
           owner_key_values = foreign_key.map { |fk| owner.read_attribute(fk) }
 
           return if !force && owner_key_values == target_key_values
@@ -172,10 +178,16 @@ module ActiveRecord
         end
 
         def stale_state
-          values = foreign_key.map do |fk|
-            owner.read_attribute(fk) { |n| owner.send(:missing_attribute, n, caller) }
+          keys = if reflection.options[:query_constraints]
+            reflection.foreign_key
+          else
+            reflection.query_foreign_key
           end
-          foreign_key.composite? ? (values if values.any?) : values.first
+          key = ActiveRecord::Key.for(keys)
+          values = key.map do |column|
+            owner.read_attribute(column) { |name| owner.send(:missing_attribute, name, caller) }
+          end
+          key.composite? ? (values if values.any?) : values.first
         end
     end
   end