+++ activerecord/lib/active_record/autosave_association.rb
@@ -492,12 +492,11 @@ module ActiveRecord
           return unless (autosave && record.changed_for_autosave?) || _record_changed?(reflection, record, primary_key_value)
 
           unless reflection.through_reflection
-            foreign_key = ActiveRecord::Key.for(reflection.foreign_key)
-            primary_key_foreign_key_pairs = primary_key.zip(foreign_key)
-
-            primary_key_foreign_key_pairs.each do |primary_key, foreign_key|
-              association_id = read_attribute(primary_key)
-              record.write_attribute(foreign_key, association_id) unless record.read_attribute(foreign_key) == association_id
+            reflection.query_key_mapping.foreign_key_pairs.each do |active_record_column, associated_record_column|
+              association_id = read_attribute(active_record_column)
+              if record.read_attribute(associated_record_column) != association_id
+                record.write_attribute(associated_record_column, association_id)
+              end
             end
             association.set_inverse_instance(record)
           end
@@ -544,9 +543,11 @@ module ActiveRecord
       # In addition, it will destroy the association if it was marked for destruction.
       def save_belongs_to_association(reflection)
         association = association_instance_get(reflection.name)
-        return unless association && association.loaded? && !association.stale_target?
+        return unless association && association.loaded?
+        stale_target = association.stale_target?
+        return if stale_target && !association.target&.new_record?
 
-        record = association.load_target
+        record = stale_target ? association.target : association.load_target
         if record && !record.destroyed?
           autosave = reflection.options[:autosave]
 
@@ -566,13 +567,9 @@ module ActiveRecord
             end
 
             if association.updated?
-              primary_key = ActiveRecord::Key.for(reflection.association_primary_key(record.class))
-              foreign_key = ActiveRecord::Key.for(reflection.foreign_key)
-
-              primary_key_foreign_key_pairs = primary_key.zip(foreign_key)
-              primary_key_foreign_key_pairs.each do |primary_key, foreign_key|
-                association_id = record.read_attribute(primary_key)
-                write_attribute(foreign_key, association_id) unless read_attribute(foreign_key) == association_id
+              reflection.query_key_mapping(record.class).foreign_key_pairs.each do |active_record_column, associated_record_column|
+                association_id = record.read_attribute(associated_record_column)
+                write_attribute(active_record_column, association_id) unless read_attribute(active_record_column) == association_id
               end
               association.loaded!
             end